Fix typo and move hook inside context file
